app/requirements.txt
flask==2.3.2
uvicorn==0.22.0
gunicorn==20.1.0
requests==2.31.0
python-dotenv==1.0.0
# LLM / agent libs (verificar nomes/versões reais)
crewai==0.1.0
langchain==0.0.300
langgraph==0.1.0
modelcontextprotocol==0.1.0
openai==1.0.0
faiss-cpu==1.7.4